fix: 3 silent bug — PG13'te pg_stat_wal yok + round(double,int) hata + idle_in_tx SQL
Bug 1 — Pg13Queries.walQuery() pg_stat_wal sorguluyordu ama bu view
PG14'te eklendi. PG13 instance'larinda her cycle'da:
  ERROR: relation "pg_stat_wal" does not exist
loglara dusuyor, cluster WAL metrikleri toplanmiyordu.
Cozum: Pg13Queries.walQuery() override silindi (superclass null doner,
collector da null-safe). Pg14_16Queries'e walQuery override eklendi —
PG14+ instance'lar pg_stat_wal'i okur.

Bug 2 — round(double precision, integer) PostgreSQL'de YOK; sadece
round(numeric, integer) var. 4 yerde bu pattern vardi:
  - AlertEvidenceResolver.resolveHighConnectionUsage (usage_pct)
  - ActionableAlertEvaluator.checkHighConnectionUsage (usage_pct)
  - AlertRuleEvaluator threshold change_pct hesabi (3 yer)
Hepsinde 100.0 (double) * value / nullif(prev_val, 0) ifadesi double
doner, round 2-arg int ile cagrildiginda hata. Cozum: tum bolum
ifadesini ::numeric ile cast.

Bug 3 — checkIdleInTxTimeHigh sorgusunda
  nullif(sum(d.session_time_ms_delta), 0)
session_time_ms_delta double precision oldugu icin nullif sonucu double,
sonraki round 2-arg int ile patliyordu. Cozum: ::numeric cast eklendi.

+++ collector/src/main/java/com/pgstat/collector/service/ActionableAlertEvaluator.java
@@ -684,7 +684,7 @@ public class ActionableAlertEvaluator {
                    sum(d.idle_in_transaction_time_ms_delta) as idle_ms,
                    sum(d.session_time_ms_delta) as session_ms,
                    round(100.0 * sum(d.idle_in_transaction_time_ms_delta)::numeric /
-                         nullif(sum(d.session_time_ms_delta), 0), 1) as idle_pct,
+                         nullif(sum(d.session_time_ms_delta)::numeric, 0), 1) as idle_pct,
                    i.display_name
             from fact.pg_database_delta d
             join control.instance_inventory i on i.instance_pk = d.instance_pk
@@ -898,7 +898,7 @@ public class ActionableAlertEvaluator {
             )
             select c.instance_pk, i.display_name, c.total_backends,
                    m.max_connections,
-                   round(100.0 * c.total_backends / nullif(m.max_connections, 0), 1) as usage_pct
+                   round((100.0 * c.total_backends / nullif(m.max_connections, 0))::numeric, 1) as usage_pct
             from conn_count c
             join control.instance_inventory i on i.instance_pk = c.instance_pk
             join max_conn m on m.instance_pk = c.instance_pk
